Sort by most stale and slice at the end
